Refactor project structure for docker compose. Directory structure complete.
