+++ backend/package.json
@@ -1,0 +1,15 @@
+{
+  "scripts": {
+    "start": "npx ts-node ./src/index.ts",
+    "build": "npx tsc && npm run build-frontend",
+    "dev-server": "tsc-watch --noClear -p ./tsconfig.json --onSuccess \"node ./dist/index.js\""
+  },
+  "dependencies": {
+    "@types/express": "^4.17.21",
+    "express": "^4.19.2",
+    "typescript": "^5.4.5"
+  },
+  "devDependencies": {
+    "tsc-watch": "^6.2.0"
+  }
+}
