+++ frontend/README.md
@@ -1,0 +1,1 @@
+# Invent Frontend
