--- package.json
@@ -1,14 +1,0 @@
-{
-  "scripts": {
-    "start": "npx ts-node index.ts",
-    "build": "npx tsc && npm run build-frontend",
-    "build-frontend": "cd frontend && npm run build-windows",
-    "install-core": "npm i",
-    "install-all": "pushd frontend && npm i && popd && npm i"
-  },
-  "dependencies": {
-    "@types/express": "^4.17.21",
-    "express": "^4.19.2",
-    "typescript": "^5.4.5"
-  }
-}
